pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
]
